boxed byte mapping

+++ example/src/main/kotlin/com/klee0kai/example/Main.kt
@@ -5,7 +5,7 @@ import com.klee0kai.reflection.ReflectionAnalyzer
 
 fun main(arg: Array<String>) {
     val simple = Simple()
-    val un: Short? = 1
+    val un: Byte? = 1
     ReflectionAnalyzer.analyze(un)
 
 //    println("init ${SimpleJniEngine.initLib()}")
